refactor: move camera factories from alarm to camera app
